# Optimizing ValidateDependenciesChecker for JoinNode