The planning time of presto is bit high when compared with the previous presto version (Issue : #11024). So we profiled and figured out that `ValidateDependenciesChecker.java` is taking more to validate a IntermediatePlan. This PR optimizes `ValidateDependenciesChecker.java` so that the planning time can be reduced for queries. 

**Benchmark for initial version**
Benchmark (iterativeOptimizerEnabled) (stage) Mode Cnt Score Error Units
BenchmarkPlanner.planQueries true optimized avgt 20 431.245 ± 15.984 ms/op
BenchmarkPlanner.planQueries true created avgt 20 142.711 ± 2.069 ms/op
BenchmarkPlanner.planQueries false optimized avgt 20 437.861 ± 27.142 ms/op
BenchmarkPlanner.planQueries false created avgt 20 142.856 ± 2.397 ms/op

**Benchmark after optimizing the `JoinNode` validation**
Benchmark (iterativeOptimizerEnabled) (stage) Mode Cnt Score Error Units
BenchmarkPlanner.planQueries true optimized avgt 20 313.418 ± 16.652 ms/op
BenchmarkPlanner.planQueries true created avgt 20 14.891 ± 1.257 ms/op
BenchmarkPlanner.planQueries false optimized avgt 20 298.709 ± 11.603 ms/op
BenchmarkPlanner.planQueries false created avgt 20 14.491 ± 0.460 ms/op