@atris 2018-07-22
Please perform some wordsmithing on the commit message. It is too verbose to get a gist of changes