@Praveen2112 2018-07-29
@kokosing Updated the code and also ran the `BenchmarkPlanner` for benchmarking.
For the initial version : 
Benchmark                     (iterativeOptimizerEnabled)    (stage)  Mode  Cnt     Score     Error  Units
BenchmarkPlanner.planQueries                         true  optimized  avgt   20  2199.486 ± 316.666  ms/op
BenchmarkPlanner.planQueries                         true    created  avgt   20   150.316 ±  16.532  ms/op
BenchmarkPlanner.planQueries                        false  optimized  avgt   20  1687.298 ± 215.590  ms/op
BenchmarkPlanner.planQueries                        false    created  avgt   20    79.258 ±   6.999  ms/op

After applying the patch:
Benchmark                     (iterativeOptimizerEnabled)    (stage)  Mode  Cnt     Score     Error  Units
BenchmarkPlanner.planQueries                         true  optimized  avgt   20  1258.897 ±  50.254  ms/op
BenchmarkPlanner.planQueries                         true    created  avgt   20    70.772 ±   6.227  ms/op
BenchmarkPlanner.planQueries                        false  optimized  avgt   20   933.292 ± 208.551  ms/op
BenchmarkPlanner.planQueries                        false    created  avgt   20    72.930 ±  11.245  ms/op