@Praveen2112 2018-07-30
@kokosing Handled those minor changes.

**Benchmark for the initial version** 
BenchmarkPlanner.planQueries                         true  optimized  avgt   20  1542.007 ± 124.085  ms/op
BenchmarkPlanner.planQueries                         true    created  avgt   20    77.899 ±   6.070  ms/op
BenchmarkPlanner.planQueries                        false  optimized  avgt   20   952.575 ± 117.280  ms/op
BenchmarkPlanner.planQueries                        false    created  avgt   20    81.141 ±   4.798  ms/op

**Benchmark after the change**
BenchmarkPlanner.planQueries                         true  optimized  avgt   20  1432.079 ± 120.257  ms/op
BenchmarkPlanner.planQueries                         true    created  avgt   20    74.664 ±   4.194  ms/op
BenchmarkPlanner.planQueries                        false  optimized  avgt   20   892.081 ±  89.001  ms/op
BenchmarkPlanner.planQueries                        false    created  avgt   20    79.808 ±   6.703  ms/op


Though not much of a difference is seen the tpch queries, for queries involving wide tables and more number of joins, the time taken for validation reduces from **1400ms** to **70ms**