@kokosing 2018-07-30
> Though not much of a difference is seen the tpch queries, for queries involving wide tables and more number of joins, the time taken for validation reduces from 1400ms to 70ms

Can you please provide a benchmark for this case. This case seems to be very important, having such benchmark will be useful for further optimizations.