@kokosing 2018-07-30
Regarding the benchmark numbers, now they are in line with my intuition for current benchmark.