@Praveen2112 2018-07-30
@kokosing I took a benchmark on tpch query 5 with modified tables (Number of columns in each table is increased by 50x). 

**Benchmark for initial version**  
Benchmark                     (iterativeOptimizerEnabled)    (stage)  Mode  Cnt    Score    Error  Units
BenchmarkPlanner.planQueries                         true  optimized  avgt   20  431.245 ± 15.984  ms/op
BenchmarkPlanner.planQueries                         true    created  avgt   20  142.711 ±  2.069  ms/op
BenchmarkPlanner.planQueries                        false  optimized  avgt   20  437.861 ± 27.142  ms/op
BenchmarkPlanner.planQueries                        false    created  avgt   20  142.856 ±  2.397  ms/op


**Benchmark after applying patch**
Benchmark                     (iterativeOptimizerEnabled)    (stage)  Mode  Cnt    Score    Error  Units
BenchmarkPlanner.planQueries                         true  optimized  avgt   20  313.418 ± 16.652  ms/op
BenchmarkPlanner.planQueries                         true    created  avgt   20   14.891 ±  1.257  ms/op
BenchmarkPlanner.planQueries                        false  optimized  avgt   20  298.709 ± 11.603  ms/op
BenchmarkPlanner.planQueries                        false    created  avgt   20   14.491 ±  0.460  ms/op


From these numbers we can infer that we have optimized the `ValidateDependenciesChecker` for `JoinNode` when we are validating a intermediate plan (437ms is reduced to 15ms)