@kokosing 2018-07-30
> (437ms is reduced to 15ms)

How did you calculate that?

For example from the below:
> Benchmark for initial version
Benchmark (iterativeOptimizerEnabled) (stage) Mode Cnt Score Error Units
BenchmarkPlanner.planQueries true optimized avgt 20 431.245 ± 15.984 ms/op
Benchmark after applying patch
Benchmark (iterativeOptimizerEnabled) (stage) Mode Cnt Score Error Units
BenchmarkPlanner.planQueries true optimized avgt 20 313.418 ± 16.652 ms/op

I see that benchmark is faster about 120ms. Still this is nice improvement.