@kokosing 2018-07-30
Please update the commit message, so it contains the benchmark results and it is ready to go.